Updated version

+++ modules/bulk_order/index.php
@@ -14,56 +14,67 @@
 		if (!$conn) {
 			die("Error in connection: " . pg_last_error());
 		}
-		/*//id submit is executed
-		if(isset($_POST['submit'])){
-			$id = $_POST['id'];
-			//connect
-			$conn = pg_connect("host=localhost port=5432 dbname=ahlscake user=postgres password=admin");
-			if (!$conn) {
-				die("Error in connection: " . pg_last_error());
-			}
-			//execute query then close connection
-			pg_query($conn, "INSERT INTO product (prod_id, prod_name, prod_desc, prod_img, prod_price) VALUES('{$id}','{$name}', '{$description}', '{$name}', '{$price}')");
-			pg_close($conn);
-		}*/
 	?>
 	<body>
 		<div class="wrapper">
 			<div class="main">
-				<div class="header">
-					<div class="banner-1">&nbsp;</div>
-				</div>
-				
+				<div class="header">Ahl's Cakes</div>
+
 				<div class="mid">
 					<div class="mid-left">
-						<h2 class="gap-2">Menu</h2>
-						<ul class="left-nav">
-						  <li><a href="../../index.html">Home</a></li>
-						  <li><a href="../view_all_products/">View Products</a></li>
-						  <li><a href="#">Order Online</a></li>
+						<h2 class="gap-2">MENU</h2>
+						<ul class="left-nav bmenu">
+						  <li class="top"><a href="../../">Home</a></li>
+						  <li><a href="../customer_view_products/?offset=0&pgnum=0">View Products</a></li>
+						  <li><a href="../bulk_order/">Order Online</a></li>
+						  <li><a href="../comments/">Comment</a></li>
 						  <li><a href="#">Contact</a></li>
-						  <li><a href="#">About Us</a></li>
+						  <li class="bottom"><a href="#">About Us</a></li>
 						</ul>
-						<h2 class="gap-2">Employee</h2>
-						<ul class="left-nav">
-						  <li><a href="../login/">Login</a></li>
-						  <li><a href="../sign_up/">Sign Up</a></li>
-						  <li><a href="../add_product/">Add Product</a></li>
-						  <li><a href="../delete_product/">Delete Product</a></li>
-						</ul>
+						<?php
+							session_start();
+							$top=0;
+							if(!isset($_SESSION["role"])){
+								echo "<h2 class=\"gap-2\">LOGIN</h2>
+									<ul class=\"left-nav bmenu\">
+									<li class='login top'>";
+								include("../login/index.php");
+								echo "<li class='bottom'><a href=\"../sign_up/\">Sign Up</a></li>";
+								echo "</li>
+									</ul>";
+							}
+							else{
+								$role=strtoupper($_SESSION["role"]); 
+								echo "<h2 class=\"gap-2\">{$role}</h2>
+								<ul class=\"left-nav bmenu\">";
+								if($_SESSION["role"]=="owner"){
+									echo "<li class='top'><a href=\"../add_employee/\">Add Employee</a></li>
+										  <li><a href='../delete_employee/'>Delete Employee</a></li>
+										  <li><a href='../search_employee/'>Search Employee</a></li>
+										  <li><a href='../view_all_employee/'>View All Employee</a></li>";
+									$top=1;
+								}
+								if($_SESSION["role"]=="employee" || $_SESSION["role"]=="owner"){
+										if($top==1) echo "<li><a href='../cashier_system/'>Cashier System</a></li>";
+										else echo "<li class='top'><a href='../cashier_system/'>Cashier System</a></li>";
+										echo "<li><a href='../inventory_system/'>Inventory System</a></li>
+											  <li><a href='../income_reports/'>Income Reports</a></li>
+											  <li><a href='../income_graphs/'>Income Graphs</a></li>
+											  <li><a href='../add_product/'>Add Product</a></li>
+											  <li><a href='../delete_product/'>Delete Product</a></li>
+											  <li class='bottom'><a href=\"../logout/index.php\">Logout</a></li>";
+								}
+							}
+						?>
 					</div>
 					<div class="mid-right">
 						<h1 class="gap-1">Bulk Order</h1>
-						<form name="bulkForm" method="post" action="<?php echo $_SERVER['PHP_SELF']; ?>">
-						<strong><?php
-							$today = date("Y-m-d");
-							PRINT "$today";
-						?>
+						<form name="bulkForm" method="post" action="">
+						<strong><?php echo date("F j, Y - l"); $today=date("Y-m-d")?>
 						<?php 
 								$db = pg_connect('host=localhost dbname=ahlscake user=postgres password=admin');
-								$product = pg_query($db, "select * from product_list order by prod_id");
+								$product = pg_query($db, "select * from product order by prod_id");
 								$i = 1;
-								
 								while($myrow = pg_fetch_assoc($product)){
 									$value = str_replace(" ","_",$myrow['prod_name']);
 									printf ("<input name='%s' type=hidden value=%d readonly />\n", $value,$myrow['prod_id']);
@@ -81,14 +92,13 @@
 							<tr>
 								<td> 
 									<select id="prodList">
-										<option> Brownies a la Mode </option>
-										<option> Cheezy Bar </option>
-										<option> Chocolate Cake </option>
-										<option> Crazy Caramel </option>
-										<option> Cream Cheese Swirl </option>
-										<option> Double Fudge Espresso Brownies </option>
-										<option> Food for the Gods </option>
-										<option> Revel Bars </option>
+										<?php
+											$product = pg_query($db, "select * from product order by prod_id");
+											while($myrow = pg_fetch_assoc($product)){
+												$value = str_replace(" ","_",$myrow['prod_name']);
+												printf ("<option>%s</option>", $myrow['prod_name']);
+											}
+										?>
 									</select>
 								</td>
 								<td><input type="button" id="productButton" value="Order" onclick="bulkorder(document.getElementById('prodList').value);" /></td>
@@ -124,19 +134,13 @@
 										else pg_query($db, "COMMIT"); 
 										$j++;
 									}
-									
-									echo "<script type='text/javascript'>showSuccessToast('Order Successful');</script>";
+									echo "<script type='text/javascript'>showSuccessToast('Order Successful <br> Redirecting to payment page ...'); setTimeout('redirect()', 4000);</script>";
 								}
 							}
-							
 							pg_close($db);
 						?>
-						
 						<input type="hidden" id="hiddenCounter" name="hiddenCounter" value="0" />
-						<table id="order">
-							
-						</table>
-						
+						<table id="order"></table>
 						<center><input type="submit" name="bulkSubmit" value="Submit" /></center>
 						</form>
 					</div>
@@ -144,7 +148,6 @@
 				<div class="footer">
 					<p class="copyright">© COPYRIGHT 2013 ALL RIGHTS RESERVED</p>
 				</div>
-				<?php if(isset($_POST['submit'])) echo "<script type='text/javascript'>showSuccessToast('Successfully Saved');</script>";?>
 			</div>
 		</div>
 	</body>
