+++ modules/edit_employee/index.php
@@ -33,34 +33,54 @@
 	<body>
 		<div class="wrapper">
 			<div class="main">
-				<div class="header">
-					Ahl's Cakes
-				</div>
-				
+				<div class="header">Ahl's Cakes</div>
+
 				<div class="mid">
 					<div class="mid-left">
-						<h2 class="gap-2">Menu</h2>
+						<h2 class="gap-2">MENU</h2>
 						<ul class="left-nav bmenu">
-						  <li class="top"><a href="../../index.html">Home</a></li>
-						  <li><a href="../view_all_products/">View Products</a></li>
-						  <li><a href="#">Order Online</a></li>
+						  <li class="top"><a href="../../">Home</a></li>
+						  <li><a href="../customer_view_products/?offset=0&pgnum=0">View Products</a></li>
+						  <li><a href="../bulk_order/">Order Online</a></li>
+						  <li><a href="../comments/">Comment</a></li>
 						  <li><a href="#">Contact</a></li>
 						  <li class="bottom"><a href="#">About Us</a></li>
 						</ul>
-						<h2 class="gap-2">Employee</h2>
-						<ul class="left-nav bmenu">
-						  <li class="top"><a href="../login/">Login</a></li>
-						  <li><a href="../sign_up/">Sign Up</a></li>
-						  <li><a href="../add_employee/">Add Employee</a></li>
-						  <li><a href="../delete_employee/">Delete Employee</a></li>
-						  <li><a href="../search_employee/">Search Employee</a></li>
-						  <li><a href="../view_all_employee/">View All Employee</a></li>
-						  <li><a href="../income_graphs/">Income Graphs</a></li>
-						  <li><a href="../income_reports/">Income Reports</a></li>
-						  <li><a href="../inventory_system/">Inventory System</a></li>
-						  <li><a href="../add_product/">Add Product</a></li>
-						  <li class="bottom"><a href="../delete_product/">Delete Product</a></li>
-						</ul>
+						<?php
+							session_start();
+							$top=0;
+							if(!isset($_SESSION["role"])){
+								echo "<h2 class=\"gap-2\">LOGIN</h2>
+									<ul class=\"left-nav bmenu\">
+									<li class='login top'>";
+								include("../login/index.php");
+								echo "<li class='bottom'><a href=\"../sign_up/\">Sign Up</a></li>";
+								echo "</li>
+									</ul>";
+							}
+							else{
+								$role=strtoupper($_SESSION["role"]); 
+								echo "<h2 class=\"gap-2\">{$role}</h2>
+								<ul class=\"left-nav bmenu\">";
+								if($_SESSION["role"]=="owner"){
+									echo "<li class='top'><a href=\"../add_employee/\">Add Employee</a></li>
+										  <li><a href='../delete_employee/'>Delete Employee</a></li>
+										  <li><a href='../search_employee/'>Search Employee</a></li>
+										  <li><a href='../view_all_employee/'>View All Employee</a></li>";
+									$top=1;
+								}
+								if($_SESSION["role"]=="employee" || $_SESSION["role"]=="owner"){
+										if($top==1) echo "<li><a href='../cashier_system/'>Cashier System</a></li>";
+										else echo "<li class='top'><a href='../cashier_system/'>Cashier System</a></li>";
+										echo "<li><a href='../inventory_system/'>Inventory System</a></li>
+											  <li><a href='../income_reports/'>Income Reports</a></li>
+											  <li><a href='../income_graphs/'>Income Graphs</a></li>
+											  <li><a href='../add_product/'>Add Product</a></li>
+											  <li><a href='../delete_product/'>Delete Product</a></li>
+											  <li class='bottom'><a href=\"../logout/index.php\">Logout</a></li>";
+								}
+							}
+						?>
 					</div>
 					<div class="mid-right">
 						<h1 class="gap-1">Edit Employee</h1>
